+++ src/main/java/com/divudi/bean/pharmacy/TransferIssueController.java
@@ -159,11 +159,11 @@ public class TransferIssueController implements Serializable {
                 //     //System.err.println("List "+bItem.getTmpSuggession());
                 PharmaceuticalBillItem phItem = new PharmaceuticalBillItem();
                 phItem.setBillItem(bItem);
-                phItem.setQtyInUnit((float) sq.getQty());
+                phItem.setQtyInUnit((double) sq.getQty());
                 System.err.println("Pharmac Item QTY " + phItem.getQtyInUnit());
                 phItem.setFreeQtyInUnit(i.getFreeQtyInUnit());
-                phItem.setPurchaseRateInUnit((float) sq.getStock().getItemBatch().getPurcahseRate());
-                phItem.setRetailRateInUnit((float) sq.getStock().getItemBatch().getRetailsaleRate());
+                phItem.setPurchaseRateInUnit((double) sq.getStock().getItemBatch().getPurcahseRate());
+                phItem.setRetailRateInUnit((double) sq.getStock().getItemBatch().getRetailsaleRate());
                 phItem.setStock(sq.getStock());
                 phItem.setDoe(sq.getStock().getItemBatch().getDateOfExpire());
                 phItem.setItemBatch(sq.getStock().getItemBatch());
